Store user data to user area

+++ src/actions/expenses.js
@@ -11,7 +11,8 @@ export const addExpense = (expense) => {
   
 
 export const startAddExpense = (expenseData = {}) => {
-  return (dispatch) => {
+  return (dispatch, getState) => {
+    const uid = getState().auth.uid
     const {
       description = '',
       note = '',
@@ -19,7 +20,7 @@ export const startAddExpense = (expenseData = {}) => {
       createdAt = 0
     } = expenseData
     const expense = {description, note, amount, createdAt:new Date(createdAt).getTime()}
-    addDoc(collection(db, "expenses"), expense).then((ref) => {
+    addDoc(collection(db, "users",`${uid}`,"expenses"), expense).then((ref) => {
           dispatch(addExpense({
             id: ref.id,
             ...expense
@@ -36,8 +37,9 @@ export const removeExpense = (id) => ({
 
 
 export const startRemoveExpense = ({ id } = {}) => {
-  return (dispatch) => {
-    deleteDoc(doc(db, "expenses", id));
+  return (dispatch, getState) => {
+    const uid = getState().auth.uid
+    deleteDoc(doc(db, "users",`${uid}`,"expenses", id));
     dispatch(removeExpense(id))
   }
 }
@@ -52,8 +54,9 @@ export const editExpense =(id, updates) => ({
 
 export const startEditExpense = (id ,updates) => {
   updates.createdAt = new Date(updates.createdAt).getTime()
-  return (dispatch) => {
-        setDoc(doc(db, "expenses", id), {
+  return (dispatch, getState) => {
+    const uid = getState().auth.uid
+      setDoc(doc(db, "users",`${uid}`,"expenses", id), {
         ...updates
       });
       dispatch(editExpense(id, updates))
@@ -66,8 +69,9 @@ export const setExpense = (expenses) => ({
 })
 
 export const startSetExpense = () => {
-  return (dispatch) => {
-    return getDocs(collection(db, "expenses")).then((querySnapshot) => {
+  return (dispatch, getState) => {
+    const uid = getState().auth.uid
+    return getDocs(collection(db, "users",`${uid}`,"expenses")).then((querySnapshot) => {
       const expenses = []
       querySnapshot.forEach((doc) => {
         expenses.push({
